# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Exam Prep Backend API
  description: >
    Веб-приложение для эффективной подготовки к экзаменам с использованием 
    метода интервальных повторений.
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/Layzov/diplom
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      description: Create a new user account and return JWT token
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: SecurePassword123
                name:
                  type: string
                  example: John Doe
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: Login user
      description: Authenticate user with email and password, return JWT token
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: SecurePassword123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/me/profile:
    get:
      tags: [Profile]
      summary: Get user profile
      description: Get current authenticated user's profile information
      operationId: getProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  data:
                    $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [Profile]
      summary: Update user profile
      description: Update current user's profile information
      operationId: updateProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: New Name
                settings:
                  type: object
                  properties:
                    timezone:
                      type: string
                      example: UTC
                    language:
                      type: string
                      example: ru-RU
                    theme:
                      type: string
                      enum: [light, dark]
                    notifications_enabled:
                      type: boolean
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  data:
                    $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
  /api/v1/me/subjects:
    get:
      tags: [Profile]
      summary: List user's subjects
      description: Get all subjects created by current user
      operationId: listUserSubjects
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Subjects retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

    post:
      tags: [Subjects]
      summary: Create subject
      description: Create a new study subject for current user
      operationId: createSubject
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  maxLength: 255
                  example: Математика
                description:
                  type: string
                  example: Подготовка к ОГЭ
      responses:
        '201':
          description: Subject created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  data:
                    $ref: '#/components/schemas/Subject'
        '400':
          description: Invalid request
  /api/v1/subjects/{id}:
    get:
      tags: [Subjects]
      summary: Get subject
      operationId: getSubject
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Subject retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    $ref: '#/components/schemas/Subject'
        '404':
          description: Subject not found

    put:
      tags: [Subjects]
      summary: Update subject
      operationId: updateSubject
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Subject updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    $ref: '#/components/schemas/Subject'

    delete:
      tags: [Subjects]
      summary: Delete subject
      operationId: deleteSubject
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Subject deleted
        '404':
          description: Subject not found
  /api/v1/subjects/{id}/topics:
    get:
      tags: [Topics]
      summary: List topics by subject
      operationId: listTopicsBySubject
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Topics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Topic'

    post:
      tags: [Topics]
      summary: Create topic
      operationId: createTopic
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  example: Основные понятия
      responses:
        '201':
          description: Topic created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    $ref: '#/components/schemas/Topic'
  /api/v1/topics/{id}:
    get:
      tags: [Topics]
      summary: Get topic
      operationId: getTopic
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Topic retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    $ref: '#/components/schemas/Topic'

    put:
      tags: [Topics]
      summary: Update topic
      operationId: updateTopic
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
      responses:
        '200':
          description: Topic updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    $ref: '#/components/schemas/Topic'

    delete:
      tags: [Topics]
      summary: Delete topic
      operationId: deleteTopic
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Topic deleted
  /api/v1/topics/{id}/tasks:
    get:
      tags: [Tasks]
      summary: List tasks by topic
      operationId: listTasksByTopic
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tasks retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'

    post:
      tags: [Tasks]
      summary: Create task
      operationId: createTask
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content, type]
              properties:
                title:
                  type: string
                  example: Задача 1
                content:
                  type: string
                  example: Решите уравнение...
                type:
                  type: string
                  enum: [flashcard, test, short_answer, fill_in_the_blank, matching, manual_review]
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    $ref: '#/components/schemas/Task'
  /api/v1/tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get task
      operationId: getTask
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    $ref: '#/components/schemas/Task'

    put:
      tags: [Tasks]
      summary: Update task
      operationId: updateTask
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
                type:
                  type: string
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    $ref: '#/components/schemas/Task'

    delete:
      tags: [Tasks]
      summary: Delete task
      operationId: deleteTask
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Task deleted
  /api/v1/sessions:
    post:
      tags: [Sessions]
      summary: Create study session
      operationId: createSession
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subject_id:
                  type: string
                  format: uuid
                notes:
                  type: string
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    $ref: '#/components/schemas/Session'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
    AuthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
            name:
              type: string
            access_token:
              type: string
              description: JWT token for authenticated requests
            token_type:
              type: string
              example: Bearer
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        role:
          type: string
          example: user
        settings:
          type: object
          properties:
            timezone:
              type: string
            language:
              type: string
            theme:
              type: string
            notifications_enabled:
              type: boolean
        created_at:
          type: string
          format: date-time
...